sorted exports created helper functions for creating default managers
