+++ viewer/src/datamodels/cad/createCadManager.ts
@@ -13,41 +13,39 @@ import { CadSectorParser } from './sector/CadSectorParser';
 import { CachedRepository } from './sector/CachedRepository';
 import { SimpleAndDetailedToSector3D } from './sector/SimpleAndDetailedToSector3D';
 import { ByVisibilityGpuSectorCuller } from './sector/culling/ByVisibilityGpuSectorCuller';
-import { File3dFormat } from '@/utilities/types';
 import { LocalModelDataClient } from '@/utilities/networking/LocalModelDataClient';
 import { CdfModelDataClient } from '@/utilities/networking/CdfModelDataClient';
+import { LocalModelIdentifier, CdfModelIdentifier, ModelDataClient } from '@/utilities/networking/types';
+import { RevealOptions } from '@/public/types';
 
-export function createDefaultCadManager(client: LocalModelDataClient): CadManager<{ fileName: string }>;
-export function createDefaultCadManager(
-  client: CdfModelDataClient
-): CadManager<{ modelId: number; revisionId: number; format: File3dFormat }>;
-export function createDefaultCadManager(
-  client: LocalModelDataClient | CdfModelDataClient
-): CadManager<{ fileName: string }> | CadManager<{ modelId: number; revisionId: number; format: File3dFormat }> {
+export function createLocalCadManager(
+  client: LocalModelDataClient,
+  options: RevealOptions = {}
+): CadManager<LocalModelIdentifier> {
+  return createCadManager(client, options);
+}
+export function createCdfCadManager(
+  client: CdfModelDataClient,
+  options: RevealOptions = {}
+): CadManager<CdfModelIdentifier> {
+  return createCadManager(client, options);
+}
+
+export function createCadManager<T>(client: ModelDataClient<T>, options: RevealOptions): CadManager<T> {
   const cadModelTransformationProvider = new DefaultCadTransformation();
   const cadMetadataParser = new CadMetadataParser();
-  let cadModelMetadataRepository: CadModelMetadataRepository<unknown>;
-  if (client instanceof CdfModelDataClient) {
-    cadModelMetadataRepository = new CadModelMetadataRepository(
-      client,
-      cadModelTransformationProvider,
-      cadMetadataParser
-    );
-  } else if (client instanceof LocalModelDataClient) {
-    cadModelMetadataRepository = new CadModelMetadataRepository(
-      client,
-      cadModelTransformationProvider,
-      cadMetadataParser
-    );
-  } else {
-    throw new Error('Client type not supported');
-  }
+  const cadModelMetadataRepository = new CadModelMetadataRepository(
+    client,
+    cadModelTransformationProvider,
+    cadMetadataParser
+  );
   const materialManager = new MaterialManager();
   const cadModelFactory = new CadModelFactory(materialManager);
   const modelDataParser = new CadSectorParser();
   const modelDataTransformer = new SimpleAndDetailedToSector3D(materialManager);
   const cachedSectorRepository = new CachedRepository(client, modelDataParser, modelDataTransformer);
-  const sectorCuller = new ByVisibilityGpuSectorCuller();
+  const { internal } = options;
+  const sectorCuller = internal && internal.sectorCuller ? internal.sectorCuller : new ByVisibilityGpuSectorCuller();
   const cadModelUpdateHandler = new CadModelUpdateHandler(cachedSectorRepository, sectorCuller);
   return new CadManager(materialManager, cadModelMetadataRepository, cadModelFactory, cadModelUpdateHandler);
 }
