+++ viewer/src/experimental.ts
@@ -5,7 +5,7 @@
 // Everything that is exposed to the user should be defined here
 
 export { RevealManager } from './public/RevealManager';
-export { LocalHostRevealManager } from './public/LocalHostRevealManager';
+export { createLocalRevealManager, createCdfRevealManager } from './public/createRevealManager';
 export { RenderManager } from './public/RenderManager';
 // TODO 2020-05-15 larsmoa: reveal.CadNode is _internal_ and should not be exported
 export { CadNode } from './datamodels/cad';
