+++ viewer/src/public/migration/Cognite3DViewer.ts
@@ -9,7 +9,7 @@ import omit from 'lodash/omit';
 import ComboControls from '@cognite/three-combo-controls';
 import { CogniteClient } from '@cognite/sdk';
 import { debounceTime, filter, map, publish } from 'rxjs/operators';
-import { merge, Subject, Subscription } from 'rxjs';
+import { merge, Subject, Subscription, fromEventPattern } from 'rxjs';
 
 import { from3DPositionToRelativeViewportCoordinates } from '@/utilities/worldToViewport';
 import { intersectCadNodes } from '@/datamodels/cad/picking';
@@ -27,7 +27,10 @@ import { BoundingBoxClipper, File3dFormat, isMobileOrTablet } from '@/utilities'
 import { Spinner } from '@/utilities/Spinner';
 import { addPostRenderEffects } from '@/datamodels/cad/rendering/postRenderEffects';
 import { trackError, initMetrics, trackLoadModel } from '@/utilities/metrics';
-import { CdfModelManager } from './CdfModelManager';
+import { RevealManager } from '../RevealManager';
+import { createCdfRevealManager } from '../createRevealManager';
+import { CdfModelIdentifier } from '@/utilities/networking/types';
+import { SectorNodeIdToTreeIndexMapLoadedEvent } from '../types';
 
 type PointerEventDelegate = (event: { offsetX: number; offsetY: number }) => void;
 type CameraChangeDelegate = (position: THREE.Vector3, target: THREE.Vector3) => void;
@@ -47,8 +50,8 @@ export class Cognite3DViewer {
   private readonly controls: ComboControls;
   private readonly sdkClient: CogniteClient;
   private readonly _updateCameraNearAndFarSubject: Subject<THREE.PerspectiveCamera>;
-  private readonly _loadingSubscription = new Subscription();
-  private readonly cdfModelManager: CdfModelManager;
+  private readonly _subscription = new Subscription();
+  private readonly _revealManager: RevealManager<CdfModelIdentifier>;
 
   private readonly eventListeners = {
     cameraChange: new Array<CameraChangeDelegate>(),
@@ -140,12 +143,18 @@ export class Cognite3DViewer {
     this.sdkClient = options.sdk;
     this.renderController = new RenderController(this.camera);
 
-    this.cdfModelManager = new CdfModelManager(this.sdkClient, options);
+    const revealOptions = { internal: {} };
+    revealOptions.internal = { _sectorCuller: options._sectorCuller };
+
+    this._revealManager = createCdfRevealManager(this.sdkClient, revealOptions);
 
     this.startPointerEventListeners();
 
-    this._loadingSubscription.add(
-      this.cdfModelManager.getLoadingStateObserver().subscribe(
+    this._subscription.add(
+      fromEventPattern(
+        h => this._revealManager.on('loadingStateChanged', h),
+        h => this._revealManager.off('loadingStateChanged', h)
+      ).subscribe(
         isLoading => {
           if (isLoading) {
             this.spinner.show();
@@ -181,8 +190,8 @@ export class Cognite3DViewer {
       cancelAnimationFrame(this.latestRequestId);
     }
 
-    this._loadingSubscription.unsubscribe();
-    this.cdfModelManager.dispose();
+    this._subscription.unsubscribe();
+    this._revealManager.dispose();
     this.domElement.removeChild(this.canvas);
     this.renderer.dispose();
     this.scene.dispose();
@@ -256,12 +265,27 @@ export class Cognite3DViewer {
     }
     if (options.geometryFilter) {
       this._geometryFilters.push(options.geometryFilter);
-      this.setSlicingPlanes(this.cdfModelManager.getSlicingPlanes());
+      this.setSlicingPlanes(this._revealManager.clippingPlanes);
     }
-
-    const model3d = await this.cdfModelManager.createCadModel(options);
+    const { modelId, revisionId } = options;
+    const cadNode = await this._revealManager.addModel('cad', {
+      modelId,
+      revisionId
+    });
+    const model3d = new Cognite3DModel(modelId, revisionId, cadNode, this.sdkClient);
     this.models.push(model3d);
     this.scene.add(model3d);
+    this._subscription.add(
+      fromEventPattern<SectorNodeIdToTreeIndexMapLoadedEvent>(
+        h => this._revealManager.on('nodeIdToTreeIndexMapLoaded', h),
+        h => this._revealManager.off('nodeIdToTreeIndexMapLoaded', h)
+      ).subscribe(event => {
+        // TODO 2020-07-05 larsmoa: Fix a better way of identifying a model than blobUrl
+        if (event.blobUrl === cadNode.cadModelMetadata.blobUrl) {
+          model3d.updateNodeIdMaps(event.nodeIdToTreeIndexMap);
+        }
+      })
+    );
 
     return model3d;
   }
@@ -289,8 +313,12 @@ export class Cognite3DViewer {
     if (options.orthographicCamera) {
       throw new NotSupportedInMigrationWrapperError();
     }
-
-    const model = await this.cdfModelManager.createPointCloudModel(options.modelId, options.revisionId);
+    const { modelId, revisionId } = options;
+    const [potreeGroup, potreeNode] = await this._revealManager.addModel('pointcloud', {
+      modelId,
+      revisionId
+    });
+    const model = new CognitePointCloudModel(modelId, revisionId, potreeGroup, potreeNode);
     this.models.push(model);
     this.scene.add(model);
     return model;
@@ -343,7 +371,7 @@ export class Cognite3DViewer {
 
     const combinedSlicingPlanes = slicingPlanes.concat(geometryFilterPlanes);
     this.renderer.localClippingEnabled = combinedSlicingPlanes.length > 0;
-    this.cdfModelManager.setSlicingPlanes(combinedSlicingPlanes);
+    this._revealManager.clippingPlanes = combinedSlicingPlanes;
     this._slicingNeedsUpdate = true;
   }
 
@@ -404,7 +432,7 @@ export class Cognite3DViewer {
   }
 
   forceRerender(): void {
-    this.cdfModelManager.requestRedraw();
+    this._revealManager.requestRedraw();
   }
 
   enableKeyboardNavigation(): void {
@@ -440,7 +468,7 @@ export class Cognite3DViewer {
 
     this.renderer.setSize(width, height);
     this.renderer.render(this.scene, screenshotCamera);
-    addPostRenderEffects(this.cdfModelManager.revealManager, this.renderer, screenshotCamera, this.scene);
+    addPostRenderEffects(this._revealManager, this.renderer, screenshotCamera, this.scene);
     const url = this.renderer.domElement.toDataURL();
 
     this.renderer.setSize(originalWidth, originalHeight);
@@ -595,19 +623,19 @@ export class Cognite3DViewer {
       }
       this.controls.update(this.clock.getDelta());
       renderController.update();
-      this.cdfModelManager.update(this.camera);
+      this._revealManager.update(this.camera);
 
       if (
         renderController.needsRedraw ||
         this.forceRendering ||
-        this.cdfModelManager.needsRedraw ||
+        this._revealManager.needsRedraw ||
         this._slicingNeedsUpdate
       ) {
         this.triggerUpdateCameraNearAndFar();
         this.renderer.render(this.scene, this.camera);
-        addPostRenderEffects(this.cdfModelManager.revealManager, this.renderer, this.camera, this.scene);
+        addPostRenderEffects(this._revealManager, this.renderer, this.camera, this.scene);
         renderController.clearNeedsRedraw();
-        this.cdfModelManager.resetRedraw();
+        this._revealManager.resetRedraw();
         this._slicingNeedsUpdate = false;
       }
     }
